requirements.txt
flask
requests
config
google.generativeai